+++ crates/module-system/sov-test-utils/src/runtime/tests.rs
@@ -217,7 +217,7 @@ fn test_custom_transaction_details_max_fee() {
         assert: Box::new(move |result, _state| {
            match &result.tx_receipt {
                 sov_modules_api::TxEffect::Skipped(reason) => {
-                    if let SkippedReason::CannotReserveGas(error_message) = reason {
+                    if let SkippedReason::OutOfGas(error_message) = reason {
                         assert!(
                             error_message.contains("The gas to charge is greater than the funds available in the meter."),
                             "Error message doesn't contain with the expected phrase. Got: {}",